+++ internal/secrets/keychain_darwin_test.go
@@ -3,6 +3,7 @@
 package secrets
 
 import (
+	"context"
 	"strings"
 	"testing"
 )
@@ -42,7 +43,7 @@ func TestKeychainPath(t *testing.T) {
 
 func TestEnsureKeychainAccess_UnlockedKeychain(t *testing.T) {
 	// On a normal dev machine, keychain should be unlocked
-	err := EnsureKeychainAccess()
+	err := EnsureKeychainAccessContext(context.Background())
 	if err != nil {
 		t.Skipf("Keychain appears to be locked, skipping: %v", err)
 	}